Check validity of the given scientific or vernacular name(s)

Check names against all scientific and vernacular names contained in a camera trap data package.

check_species(datapkg, species, arg_name = "species")

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp().

species
a character vector with scientific or vernacular names

arg_name
character with argument name to return in error message Default: "species".

Value

a character vector with the correspondent scientific names

Examples

# species is a scientific name
check_species(mica, "Martes foina")
#> [1] "Martes foina"
# species is a vector of vernacular names
 check_species(mica, c("beech marten", "european polecat"))
#> Scientific name of beech marten: Martes foina
#> Scientific name of european polecat: Mustela putorius
#> [1] "Martes foina"     "Mustela putorius"
# vernacular names can be specified in any language available
check_species(mica, c("vos", "blauwe reiger"))
#> Scientific name of vos: Vulpes vulpes
#> Scientific name of blauwe reiger: Ardea cinerea
#> [1] "Vulpes vulpes" "Ardea cinerea"
# vernacular names and scientific names can be mixed up
check_species(mica, c("beech marten", "blauwe reiger", "Anas strepera"))
#> Scientific name of beech marten: Martes foina
#> Scientific name of blauwe reiger: Ardea cinerea
#> [1] "Martes foina"  "Ardea cinerea" "Anas strepera"
# case insensitive
check_species(mica, "AnaS StrePeRa")
#> [1] "Anas strepera"
check_species(mica, "bEEch mARteN")
#> Scientific name of bEEch mARteN: Martes foina
#> [1] "Martes foina"
if (FALSE) {
check_species(mica, "bad name")
}