Get camera operation matrix

This function returns the camera operation matrix as returned by camtrapR's function cameraOperation().

get_cam_op(datapkg, ...)

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp(), i.e. a list containing three data.frames:

observations

deployments

multimedia

and a list with metadata: datapackage

...
filter predicates for filtering on deployments

Value

a matrix. Row names always indicate the location_name (station ID) "Station"+location_name. Column names are dates.

Details

The deployment data are grouped by location_name (station ID in camtrapR jargon). If multiple deploymnts are linked to same location, daily efforts higher than one occur.

Examples

get_cam_op(camtrapdp)

#> Note: method with signature ‘Duration#ANY’ chosen for function ‘-’,
#>  target signature ‘Duration#Duration’.
#>  "ANY#Duration" would also be valid

#>                                    2020-06-03 2020-06-04 2020-06-05 2020-06-06
#> StationB_ML_val 03_De Val           0.1533565          1          1          1
#> StationB_ML_val 05_molenkreek       0.1595139          1          1          1
#> StationB_ML_val 06_Oostpolderkreek  0.1635069          1          1          1
#> StationB_ML_val 07_Sint-Anna        0.1160648          1          1          1
#>                                    2020-06-07 2020-06-08 2020-06-09 2020-06-10
#> StationB_ML_val 03_De Val          1.00000000 1.00000000          1          1
#> StationB_ML_val 05_molenkreek      1.00000000 1.00000000          1          1
#> StationB_ML_val 06_Oostpolderkreek 0.01864583 0.00000000          0          0
#> StationB_ML_val 07_Sint-Anna       1.00000000 0.01019676          0          0
#>                                    2020-06-11 2020-06-12 2020-06-13 2020-06-14
#> StationB_ML_val 03_De Val                   1          1          1          1
#> StationB_ML_val 05_molenkreek               1          1          1          1
#> StationB_ML_val 06_Oostpolderkreek          0          0          0          0
#> StationB_ML_val 07_Sint-Anna                0          0          0          0
#>                                    2020-06-15 2020-06-16 2020-06-17 2020-06-18
#> StationB_ML_val 03_De Val                   1          1          1          1
#> StationB_ML_val 05_molenkreek               1          1          1          1
#> StationB_ML_val 06_Oostpolderkreek          0          0          0          0
#> StationB_ML_val 07_Sint-Anna                0          0          0          0
#>                                    2020-06-19 2020-06-20 2020-06-21 2020-06-22
#> StationB_ML_val 03_De Val                   1          1          1          1
#> StationB_ML_val 05_molenkreek               1          1          1          1
#> StationB_ML_val 06_Oostpolderkreek          0          0          0          0
#> StationB_ML_val 07_Sint-Anna                0          0          0          0
#>                                    2020-06-23 2020-06-24 2020-06-25 2020-06-26
#> StationB_ML_val 03_De Val                   1  1.0000000          1          1
#> StationB_ML_val 05_molenkreek               1  0.9404282          0          0
#> StationB_ML_val 06_Oostpolderkreek          0  0.0000000          0          0
#> StationB_ML_val 07_Sint-Anna                0  0.0000000          0          0
#>                                    2020-06-27 2020-06-28 2020-06-29 2020-06-30
#> StationB_ML_val 03_De Val                   1          1          1          1
#> StationB_ML_val 05_molenkreek               0          0          0          0
#> StationB_ML_val 06_Oostpolderkreek          0          0          0          0
#> StationB_ML_val 07_Sint-Anna                0          0          0          0
#>                                    2020-07-01 2020-07-02 2020-07-03
#> StationB_ML_val 03_De Val                   1          1  0.8733565
#> StationB_ML_val 05_molenkreek               0          0  0.0000000
#> StationB_ML_val 06_Oostpolderkreek          0          0  0.0000000
#> StationB_ML_val 07_Sint-Anna                0          0  0.0000000

# applying filter(s) on deployments, e.g. deployments with latitude >= 51.28
get_cam_op(camtrapdp, pred_gte("latitude", 51.28))

#> df %>% filter((latitude >= 51.28))

#>                               2020-06-03 2020-06-04 2020-06-05 2020-06-06
#> StationB_ML_val 05_molenkreek  0.1595139          1          1          1
#>                               2020-06-07 2020-06-08 2020-06-09 2020-06-10
#> StationB_ML_val 05_molenkreek          1          1          1          1
#>                               2020-06-11 2020-06-12 2020-06-13 2020-06-14
#> StationB_ML_val 05_molenkreek          1          1          1          1
#>                               2020-06-15 2020-06-16 2020-06-17 2020-06-18
#> StationB_ML_val 05_molenkreek          1          1          1          1
#>                               2020-06-19 2020-06-20 2020-06-21 2020-06-22
#> StationB_ML_val 05_molenkreek          1          1          1          1
#>                               2020-06-23 2020-06-24
#> StationB_ML_val 05_molenkreek          1  0.9404282