Get custom effort

Function to get custom effort (deployment duration) for a custom time windows and a specific time interval such as day or month. The custom effort is also calculated over all deployments, although filtering predicates can be applied as well.

get_custom_effort(
  datapkg,
  ...,
  start = NULL,
  end = NULL,
  group_by = NULL,
  unit = "hour"
)

Arguments

datapkg

Camera trap data package object, as returned by read_camtrap_dp().

...

filter predicates

start

Start date. Default: NULL. If NULL the earliest start date among all deployments is used.

end

End date. Default: NULL. If NULL the latest end date among all deployments is used.

group_by

Character, one of "day", "week", "month", "year". The effort is calculated at the interval rate defined in group_by. Default: NULL: no grouping, i.e. the entire interval from start to end is taken into account as a whole.

unit

Character, the time unit to use while returning custom effort. One of: hour (default), day.

Value

A tibble (data.frame) with following columns:

begin: Date: begin of the interval the effort is calculated over.

effort: The effort as number.

effort_unit: Character specifying the effort_unit.

year: Numeric, the year begin belongs to. This column is not present if group_by is NULL.

month: Numeric, the month begin belongs to. This column is present if group_by is "month" or "day".

week: Numeric, the week begin belongs to. This column is present if group_by is "week".

day: Numeric, the day begin belongs to. This column is present if group_by is "day".

See also

Examples

# a global effort over the entire duration of the project (datapackage)
# measured in hours
get_custom_effort(mica)
#> # A tibble: 1 × 3
#>   begin      effort unit 
#>   <date>      <dbl> <chr>
#> 1 2019-10-09  1321. hour 

# global effort expressed in days
get_custom_effort(mica, unit = "day")
#> # A tibble: 1 × 3
#>   begin      effort unit 
#>   <date>      <dbl> <chr>
#> 1 2019-10-09   55.0 day  

# total effort from a specific start to a specific end
get_custom_effort(
  mica,
  start = as.Date("2019-12-15"), # or lubridate::as_date("2019-12-15")
  end = as.Date("2021-01-10")
)
#> # A tibble: 1 × 3
#>   begin      effort unit 
#>   <date>      <dbl> <chr>
#> 1 2019-12-15   457. hour 

# effort at daily interval
get_custom_effort(
  mica,
  group_by = "day"
)
#> # A tibble: 558 × 6
#>    begin      effort unit   year month   day
#>    <date>      <dbl> <chr> <dbl> <dbl> <int>
#>  1 2019-10-09   12.7 hour   2019    10     9
#>  2 2019-10-10   24   hour   2019    10    10
#>  3 2019-10-11   24   hour   2019    10    11
#>  4 2019-10-12   24   hour   2019    10    12
#>  5 2019-10-13   24   hour   2019    10    13
#>  6 2019-10-14   24   hour   2019    10    14
#>  7 2019-10-15   24   hour   2019    10    15
#>  8 2019-10-16   24   hour   2019    10    16
#>  9 2019-10-17   24   hour   2019    10    17
#> 10 2019-10-18   24   hour   2019    10    18
#> # … with 548 more rows
# effort at weekly interval
get_custom_effort(
  mica,
  group_by = "week"
)
#> # A tibble: 82 × 5
#>    begin      effort unit   year  week
#>    <date>      <dbl> <chr> <dbl> <dbl>
#>  1 2019-10-09  133.  hour   2019    41
#>  2 2019-10-15  168   hour   2019    42
#>  3 2019-10-22   34.0 hour   2019    43
#>  4 2019-10-29    0   hour   2019    44
#>  5 2019-11-05    0   hour   2019    45
#>  6 2019-11-12    0   hour   2019    46
#>  7 2019-11-19    0   hour   2019    47
#>  8 2019-11-26    0   hour   2019    48
#>  9 2019-12-03    0   hour   2019    49
#> 10 2019-12-10    0   hour   2019    50
#> # … with 72 more rows
# effort at monthly interval
get_custom_effort(
  mica,
  group_by = "month"
)
#> # A tibble: 19 × 5
#>    begin      effort unit   year month
#>    <date>      <dbl> <chr> <dbl> <dbl>
#>  1 2019-10-09  335.  hour   2019    10
#>  2 2019-11-01    0   hour   2019    11
#>  3 2019-12-01    0   hour   2019    12
#>  4 2020-01-01    0   hour   2020     1
#>  5 2020-02-01    0   hour   2020     2
#>  6 2020-03-01    0   hour   2020     3
#>  7 2020-04-01    0   hour   2020     4
#>  8 2020-05-01    0   hour   2020     5
#>  9 2020-06-01  219.  hour   2020     6
#> 10 2020-07-01   66.5 hour   2020     7
#> 11 2020-08-01  172.  hour   2020     8
#> 12 2020-09-01    0   hour   2020     9
#> 13 2020-10-01    0   hour   2020    10
#> 14 2020-11-01    0   hour   2020    11
#> 15 2020-12-01    0   hour   2020    12
#> 16 2021-01-01    0   hour   2021     1
#> 17 2021-02-01    0   hour   2021     2
#> 18 2021-03-01   99.4 hour   2021     3
#> 19 2021-04-01  429.  hour   2021     4
# effort at yearly interval
get_custom_effort(
  mica,
  group_by = "year"
)
#> # A tibble: 3 × 4
#>   begin      effort unit   year
#>   <date>      <dbl> <chr> <dbl>
#> 1 2019-10-09   335. hour   2019
#> 2 2020-01-01   457. hour   2020
#> 3 2021-01-01   529. hour   2021
# applying filter(s), e.g. deployments with latitude >= 51.18
get_custom_effort(mica, pred_gte("latitude", 51.18))
#> df %>% filter((latitude >= 51.18))
#> # A tibble: 1 × 3
#>   begin      effort unit 
#>   <date>      <dbl> <chr>
#> 1 2020-06-19   457. hour