Get effort

Function to get the effort (deployment duration) per deployment.

get_effort(datapkg, ..., unit = NULL)

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp().

...
filter predicates

unit
time unit to use while returning deployment effort (duration). One of:

second

minute

hour

day

month

year

NULL (default) duration objects (e.g. 2594308s (~4.29 weeks))

Value

a tibble (data.frame) with the following columns:

deploymentID deployment unique identifier

effort: a duration object (duration is a class from lubridate package)

See also

Examples

# efforts expressed as Durations
get_effort(mica)
#> # A tibble: 4 × 3
#>   deploymentID                         effort                 effort_unit
#>   <chr>                                <Duration>             <chr>      
#> 1 29b7d356-4bb4-4ec4-b792-2af5cc32efa8 859859s (~1.42 weeks)  Duration   
#> 2 577b543a-2cf1-4b23-b6d2-cda7e2eac372 786802s (~1.3 weeks)   Duration   
#> 3 62c200a9-0e03-4495-bcd8-032944f6f5a1 1903602s (~3.15 weeks) Duration   
#> 4 7ca633fa-64f8-4cfc-a628-6b0c419056d7 1204929s (~1.99 weeks) Duration   

# effort expressed as days
get_effort(mica, unit = "day")
#> # A tibble: 4 × 3
#>   deploymentID                         effort effort_unit
#>   <chr>                                 <dbl> <chr>      
#> 1 29b7d356-4bb4-4ec4-b792-2af5cc32efa8   9.95 day        
#> 2 577b543a-2cf1-4b23-b6d2-cda7e2eac372   9.11 day        
#> 3 62c200a9-0e03-4495-bcd8-032944f6f5a1  22.0  day        
#> 4 7ca633fa-64f8-4cfc-a628-6b0c419056d7  13.9  day