Get number of individuals for each deployment

Function to get the number of individuals (of a subset of species) per deployment. The number of observed individuals is stored in field count of observations.

get_n_individuals(datapkg, ..., species = "all", sex = NULL, age = NULL)

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp(), i.e. a list containing three data.frames:

observations 2. deployments 3. multimedia

and a list with metadata: datapackage

...
filter predicates for filtering on deployments

species
a character with scientific names or common names (case insensitive). If "all", default, all scientific names are automatically selected. If NULL all observations of all species are taken into account

sex
a character defining the sex class to filter on, e.g. "female" or c("male", "undefined"). If NULL, default, all observations of all sex classes are taken into account.

age
a character vector defining the age class to filter on, e.g. "adult" or c("subadult", "adult"). If NULL, default, all observations of all age classes are taken into account.

Value

a tibble (data.frame) with the following columns:

deployment_id: deployment unique identifier

scientific_name: scientific name of the species. This column is omitted if argument species = NULL

n: (integer) number of individuals

Examples

# get number of observations for each species
get_n_individuals(camtrapdp)

#> # A tibble: 20 × 3
#>    deployment_id                        scientific_name         n
#>    <chr>                                <chr>               <int>
#>  1 3b94f935-020f-49d8-bd59-1736f8a266c4 Anas platyrhynchos     87
#>  2 3b94f935-020f-49d8-bd59-1736f8a266c4 Rattus norvegicus     217
#>  3 3b94f935-020f-49d8-bd59-1736f8a266c4 Ondatra zibethicus      3
#>  4 3b94f935-020f-49d8-bd59-1736f8a266c4 Myocastor coypus        2
#>  5 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     1
#>  6 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Anas platyrhynchos     46
#>  7 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Rattus norvegicus       0
#>  8 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Ondatra zibethicus      1
#>  9 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Myocastor coypus        0
#> 10 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 11 9e84c750-e8fa-4502-ae12-cc6c75532f5a Anas platyrhynchos     10
#> 12 9e84c750-e8fa-4502-ae12-cc6c75532f5a Rattus norvegicus       0
#> 13 9e84c750-e8fa-4502-ae12-cc6c75532f5a Ondatra zibethicus      0
#> 14 9e84c750-e8fa-4502-ae12-cc6c75532f5a Myocastor coypus        0
#> 15 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 16 fff2f46e-8163-453c-9044-61fb77587f5d Anas platyrhynchos      5
#> 17 fff2f46e-8163-453c-9044-61fb77587f5d Rattus norvegicus       7
#> 18 fff2f46e-8163-453c-9044-61fb77587f5d Ondatra zibethicus      1
#> 19 fff2f46e-8163-453c-9044-61fb77587f5d Myocastor coypus        0
#> 20 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# get number of obs of all species, not identified individuals as well
get_n_individuals(camtrapdp, species = NULL)

#> # A tibble: 4 × 2
#>   deployment_id                            n
#>   <chr>                                <int>
#> 1 3b94f935-020f-49d8-bd59-1736f8a266c4   310
#> 2 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1    47
#> 3 9e84c750-e8fa-4502-ae12-cc6c75532f5a    10
#> 4 fff2f46e-8163-453c-9044-61fb77587f5d    13

# get number of observations of Gallinula chloropus
get_n_individuals(camtrapdp, species = "Gallinula chloropus")

#> There are 3 deployments with no observations: 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1, 9e84c750-e8fa-4502-ae12-cc6c75532f5a and fff2f46e-8163-453c-9044-61fb77587f5d

#> # A tibble: 4 × 3
#>   deployment_id                        scientific_name         n
#>   <chr>                                <chr>               <int>
#> 1 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     1
#> 2 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 3 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 4 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# get number of observations of Common Moorhen
get_n_individuals(camtrapdp, species = "Common Moorhen")

#> Scientific name of Common Moorhen: Gallinula chloropus

#> There are 3 deployments with no observations: 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1, 9e84c750-e8fa-4502-ae12-cc6c75532f5a and fff2f46e-8163-453c-9044-61fb77587f5d

#> # A tibble: 4 × 3
#>   deployment_id                        scientific_name         n
#>   <chr>                                <chr>               <int>
#> 1 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     1
#> 2 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 3 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 4 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# case insensitive
get_n_individuals(camtrapdp, species = "galliNULa CHloropUs")

#> There are 3 deployments with no observations: 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1, 9e84c750-e8fa-4502-ae12-cc6c75532f5a and fff2f46e-8163-453c-9044-61fb77587f5d

#> # A tibble: 4 × 3
#>   deployment_id                        scientific_name         n
#>   <chr>                                <chr>               <int>
#> 1 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     1
#> 2 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 3 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 4 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

get_n_individuals(camtrapdp, species = "cOmmon moorhEn")

#> Scientific name of cOmmon moorhEn: Gallinula chloropus

#> There are 3 deployments with no observations: 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1, 9e84c750-e8fa-4502-ae12-cc6c75532f5a and fff2f46e-8163-453c-9044-61fb77587f5d

#> # A tibble: 4 × 3
#>   deployment_id                        scientific_name         n
#>   <chr>                                <chr>               <int>
#> 1 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     1
#> 2 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 3 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 4 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# specify age
get_n_individuals(camtrapdp, age = "adult")

#> # A tibble: 20 × 3
#>    deployment_id                        scientific_name         n
#>    <chr>                                <chr>               <int>
#>  1 3b94f935-020f-49d8-bd59-1736f8a266c4 Anas platyrhynchos     57
#>  2 3b94f935-020f-49d8-bd59-1736f8a266c4 Rattus norvegicus     205
#>  3 3b94f935-020f-49d8-bd59-1736f8a266c4 Ondatra zibethicus      3
#>  4 3b94f935-020f-49d8-bd59-1736f8a266c4 Myocastor coypus        2
#>  5 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     1
#>  6 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Anas platyrhynchos     24
#>  7 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Rattus norvegicus       0
#>  8 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Ondatra zibethicus      1
#>  9 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Myocastor coypus        0
#> 10 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 11 9e84c750-e8fa-4502-ae12-cc6c75532f5a Anas platyrhynchos      2
#> 12 9e84c750-e8fa-4502-ae12-cc6c75532f5a Rattus norvegicus       0
#> 13 9e84c750-e8fa-4502-ae12-cc6c75532f5a Ondatra zibethicus      0
#> 14 9e84c750-e8fa-4502-ae12-cc6c75532f5a Myocastor coypus        0
#> 15 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 16 fff2f46e-8163-453c-9044-61fb77587f5d Anas platyrhynchos      1
#> 17 fff2f46e-8163-453c-9044-61fb77587f5d Rattus norvegicus       7
#> 18 fff2f46e-8163-453c-9044-61fb77587f5d Ondatra zibethicus      1
#> 19 fff2f46e-8163-453c-9044-61fb77587f5d Myocastor coypus        0
#> 20 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# specify sex
get_n_individuals(camtrapdp, sex = "female")

#> There are 3 deployments with no observations: 3b94f935-020f-49d8-bd59-1736f8a266c4, 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 and 9e84c750-e8fa-4502-ae12-cc6c75532f5a

#> # A tibble: 20 × 3
#>    deployment_id                        scientific_name         n
#>    <chr>                                <chr>               <int>
#>  1 3b94f935-020f-49d8-bd59-1736f8a266c4 Anas platyrhynchos      0
#>  2 3b94f935-020f-49d8-bd59-1736f8a266c4 Rattus norvegicus       0
#>  3 3b94f935-020f-49d8-bd59-1736f8a266c4 Ondatra zibethicus      0
#>  4 3b94f935-020f-49d8-bd59-1736f8a266c4 Myocastor coypus        0
#>  5 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     0
#>  6 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Anas platyrhynchos      0
#>  7 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Rattus norvegicus       0
#>  8 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Ondatra zibethicus      0
#>  9 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Myocastor coypus        0
#> 10 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 11 9e84c750-e8fa-4502-ae12-cc6c75532f5a Anas platyrhynchos      0
#> 12 9e84c750-e8fa-4502-ae12-cc6c75532f5a Rattus norvegicus       0
#> 13 9e84c750-e8fa-4502-ae12-cc6c75532f5a Ondatra zibethicus      0
#> 14 9e84c750-e8fa-4502-ae12-cc6c75532f5a Myocastor coypus        0
#> 15 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 16 fff2f46e-8163-453c-9044-61fb77587f5d Anas platyrhynchos      1
#> 17 fff2f46e-8163-453c-9044-61fb77587f5d Rattus norvegicus       0
#> 18 fff2f46e-8163-453c-9044-61fb77587f5d Ondatra zibethicus      0
#> 19 fff2f46e-8163-453c-9044-61fb77587f5d Myocastor coypus        0
#> 20 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# specify both sex and age
get_n_individuals(camtrapdp, sex = "undefined", age = "adult")

#> There are 3 deployments with no observations: 3b94f935-020f-49d8-bd59-1736f8a266c4, 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 and 9e84c750-e8fa-4502-ae12-cc6c75532f5a

#> # A tibble: 20 × 3
#>    deployment_id                        scientific_name         n
#>    <chr>                                <chr>               <int>
#>  1 3b94f935-020f-49d8-bd59-1736f8a266c4 Rattus norvegicus       0
#>  2 3b94f935-020f-49d8-bd59-1736f8a266c4 Ondatra zibethicus      0
#>  3 3b94f935-020f-49d8-bd59-1736f8a266c4 Anas platyrhynchos      0
#>  4 3b94f935-020f-49d8-bd59-1736f8a266c4 Myocastor coypus        0
#>  5 3b94f935-020f-49d8-bd59-1736f8a266c4 Gallinula chloropus     0
#>  6 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Rattus norvegicus       0
#>  7 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Ondatra zibethicus      0
#>  8 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Anas platyrhynchos      0
#>  9 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Myocastor coypus        0
#> 10 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0
#> 11 9e84c750-e8fa-4502-ae12-cc6c75532f5a Rattus norvegicus       0
#> 12 9e84c750-e8fa-4502-ae12-cc6c75532f5a Ondatra zibethicus      0
#> 13 9e84c750-e8fa-4502-ae12-cc6c75532f5a Anas platyrhynchos      0
#> 14 9e84c750-e8fa-4502-ae12-cc6c75532f5a Myocastor coypus        0
#> 15 9e84c750-e8fa-4502-ae12-cc6c75532f5a Gallinula chloropus     0
#> 16 fff2f46e-8163-453c-9044-61fb77587f5d Rattus norvegicus       4
#> 17 fff2f46e-8163-453c-9044-61fb77587f5d Ondatra zibethicus      1
#> 18 fff2f46e-8163-453c-9044-61fb77587f5d Anas platyrhynchos      0
#> 19 fff2f46e-8163-453c-9044-61fb77587f5d Myocastor coypus        0
#> 20 fff2f46e-8163-453c-9044-61fb77587f5d Gallinula chloropus     0

# applying filter(s), e.g. deployments with latitude >= 51.28
get_n_individuals(camtrapdp, pred_gte("latitude", 51.28))

#> df %>% filter((latitude >= 51.28))

#> # A tibble: 5 × 3
#>   deployment_id                        scientific_name         n
#>   <chr>                                <chr>               <int>
#> 1 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Anas platyrhynchos     46
#> 2 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Rattus norvegicus       0
#> 3 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Ondatra zibethicus      1
#> 4 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Myocastor coypus        0
#> 5 6b1199ac-e6e4-4eb4-a625-6ee8f9db09f1 Gallinula chloropus     0