Get record table

This function calculate the record table from a camera trap data package and so tabulating species records. The record table is a concept develoepd within the camtrapR package: see article https://jniedballa.github.io/camtrapR/articles/camtrapr3.html. See also the camtrapR's function documentation recordTable

get_record_table(
  datapkg,
  stationCol = "location_name",
  exclude = NULL,
  minDeltaTime = 0,
  deltaTimeComparedTo = NULL
)

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp(), i.e. a list containing three data.frames:

observations 2. deployments 3. multimedia

and a list with metadata: datapackage

stationCol
(character) name of the column containing stations. Default: "location_name"

exclude
(character) vector of species names (scientific names or vernacular names) to be excluded from the record table. Default: NULL

minDeltaTime
(integer) time difference between records of the same species at the same station to be considered independent (in minutes)

deltaTimeComparedTo
(character) "lastIndependentRecord" or "lastRecord". For two records to be considered independent, must the second one be at least minDeltaTime minutes after the last independent record of the same species (deltaTimeComparedTo = "lastIndependentRecord" ), or minDeltaTime minutes after the last record ( deltaTimeComparedTo = "lastRecord")? If minDeltaTime is 0, deltaTimeComparedTo must be NULL (deafult)

Value

A (tibble) data frame containing species records and additional information about stations, date, time and further metadata, such as filenames and directories of the images (multimedia) linked to the species records. Some more details about the columns returned:

Station: character, station names, as found in the deployment column defined in argument stationCol

Species: character, the scientific name of the observed species

DateTimeOriginal: datetime object, as found in column timestamp of observations

Date: date object, the date part of DateTimeOriginal

Time: character, the time part of DateTimeOriginal

delta.time.secs: numeric, the duration in seconds from the previous independent record of a given species at a certain location

delta.time.mins: numeric, the duration in minutes from the previous independent record of a given species at a certain location

delta.time.hours: numeric, the duration in hours from the previous independent record of a given species at a certain location

delta.time.days: numeric, the duration in days from the previous independent record of a given species at a certain location

Directory: list, file paths of the images linked to the given record, as defined in column file_path of multimedia

Filename: list, file names of the images linked to the given record, as defined in column file_name of multimedia

Examples

get_record_table(camtrapdp)

#> # A tibble: 267 × 11
#>    Station            Species DateTimeOriginal    Date       Time  delta.time.secs
#>    <chr>              <chr>   <dttm>              <date>     <chr>           <dbl>
#>  1 B_ML_val 03_De Val Anas p… 2020-06-03 20:27:31 2020-06-03 20:2…               0
#>  2 B_ML_val 03_De Val Anas p… 2020-06-08 20:22:13 2020-06-08 20:2…          431682
#>  3 B_ML_val 03_De Val Anas p… 2020-06-10 20:30:59 2020-06-10 20:3…          173326
#>  4 B_ML_val 03_De Val Anas p… 2020-06-11 20:36:45 2020-06-11 20:3…           86746
#>  5 B_ML_val 03_De Val Anas p… 2020-06-11 20:47:05 2020-06-11 20:4…             620
#>  6 B_ML_val 03_De Val Anas p… 2020-06-18 21:07:48 2020-06-18 21:0…          606043
#>  7 B_ML_val 03_De Val Anas p… 2020-06-22 20:17:01 2020-06-22 20:1…          342553
#>  8 B_ML_val 03_De Val Anas p… 2020-06-23 03:29:34 2020-06-23 03:2…           25953
#>  9 B_ML_val 03_De Val Anas p… 2020-06-27 03:28:59 2020-06-27 03:2…          345565
#> 10 B_ML_val 03_De Val Anas p… 2020-06-27 03:40:16 2020-06-27 03:4…             677
#> # … with 257 more rows, and 5 more variables: delta.time.mins <dbl>,
#> #   delta.time.hours <dbl>, delta.time.days <dbl>, Directory <list>,
#> #   FileName <list>

# set a minDeltaTime of 20 minutes from last independent record for filtering
# out not independent observations
get_record_table(camtrapdp,
    minDeltaTime = 20,
    deltaTimeComparedTo = "lastIndependentRecord")

#> Number of not independent observations to be removed: 78

#> # A tibble: 189 × 11
#>    Station            Species DateTimeOriginal    Date       Time  delta.time.secs
#>    <chr>              <chr>   <dttm>              <date>     <chr>           <dbl>
#>  1 B_ML_val 03_De Val Anas p… 2020-06-03 20:27:31 2020-06-03 20:2…               0
#>  2 B_ML_val 03_De Val Anas p… 2020-06-08 20:22:13 2020-06-08 20:2…          431682
#>  3 B_ML_val 03_De Val Anas p… 2020-06-10 20:30:59 2020-06-10 20:3…          173326
#>  4 B_ML_val 03_De Val Anas p… 2020-06-11 20:36:45 2020-06-11 20:3…           86746
#>  5 B_ML_val 03_De Val Anas p… 2020-06-18 21:07:48 2020-06-18 21:0…          606663
#>  6 B_ML_val 03_De Val Anas p… 2020-06-22 20:17:01 2020-06-22 20:1…          342553
#>  7 B_ML_val 03_De Val Anas p… 2020-06-23 03:29:34 2020-06-23 03:2…           25953
#>  8 B_ML_val 03_De Val Anas p… 2020-06-27 03:28:59 2020-06-27 03:2…          345565
#>  9 B_ML_val 03_De Val Anas p… 2020-06-27 03:56:46 2020-06-27 03:5…            1667
#> 10 B_ML_val 03_De Val Anas p… 2020-06-27 20:20:38 2020-06-27 20:2…           59032
#> # … with 179 more rows, and 5 more variables: delta.time.mins <dbl>,
#> #   delta.time.hours <dbl>, delta.time.days <dbl>, Directory <list>,
#> #   FileName <list>

# set a minDeltaTime of 20 minutes from last record for filtering out not
# independent observations
get_record_table(camtrapdp,
    minDeltaTime = 20,
    deltaTimeComparedTo = "lastRecord")

#> Number of not independent observations to be removed: 82

#> # A tibble: 185 × 11
#>    Station            Species DateTimeOriginal    Date       Time  delta.time.secs
#>    <chr>              <chr>   <dttm>              <date>     <chr>           <dbl>
#>  1 B_ML_val 03_De Val Anas p… 2020-06-03 20:27:31 2020-06-03 20:2…               0
#>  2 B_ML_val 03_De Val Anas p… 2020-06-08 20:22:13 2020-06-08 20:2…          431682
#>  3 B_ML_val 03_De Val Anas p… 2020-06-10 20:30:59 2020-06-10 20:3…          173326
#>  4 B_ML_val 03_De Val Anas p… 2020-06-11 20:36:45 2020-06-11 20:3…           86746
#>  5 B_ML_val 03_De Val Anas p… 2020-06-18 21:07:48 2020-06-18 21:0…          606663
#>  6 B_ML_val 03_De Val Anas p… 2020-06-22 20:17:01 2020-06-22 20:1…          342553
#>  7 B_ML_val 03_De Val Anas p… 2020-06-23 03:29:34 2020-06-23 03:2…           25953
#>  8 B_ML_val 03_De Val Anas p… 2020-06-27 03:28:59 2020-06-27 03:2…          345565
#>  9 B_ML_val 03_De Val Anas p… 2020-06-27 03:56:46 2020-06-27 03:5…            1667
#> 10 B_ML_val 03_De Val Anas p… 2020-06-27 20:20:38 2020-06-27 20:2…           59032
#> # … with 175 more rows, and 5 more variables: delta.time.mins <dbl>,
#> #   delta.time.hours <dbl>, delta.time.days <dbl>, Directory <list>,
#> #   FileName <list>

# exclude observations of Norway Rat
# exclude is case insensitive and vernacular names allowed
get_record_table(camtrapdp, exclude = "Norway raT")

#> Scientific name of Norway raT: Rattus norvegicus

#> # A tibble: 59 × 11
#>    Station            Species DateTimeOriginal    Date       Time  delta.time.secs
#>    <chr>              <chr>   <dttm>              <date>     <chr>           <dbl>
#>  1 B_ML_val 03_De Val Anas p… 2020-06-03 20:27:31 2020-06-03 20:2…               0
#>  2 B_ML_val 03_De Val Anas p… 2020-06-08 20:22:13 2020-06-08 20:2…          431682
#>  3 B_ML_val 03_De Val Anas p… 2020-06-10 20:30:59 2020-06-10 20:3…          173326
#>  4 B_ML_val 03_De Val Anas p… 2020-06-11 20:36:45 2020-06-11 20:3…           86746
#>  5 B_ML_val 03_De Val Anas p… 2020-06-11 20:47:05 2020-06-11 20:4…             620
#>  6 B_ML_val 03_De Val Anas p… 2020-06-18 21:07:48 2020-06-18 21:0…          606043
#>  7 B_ML_val 03_De Val Anas p… 2020-06-22 20:17:01 2020-06-22 20:1…          342553
#>  8 B_ML_val 03_De Val Anas p… 2020-06-23 03:29:34 2020-06-23 03:2…           25953
#>  9 B_ML_val 03_De Val Anas p… 2020-06-27 03:28:59 2020-06-27 03:2…          345565
#> 10 B_ML_val 03_De Val Anas p… 2020-06-27 03:40:16 2020-06-27 03:4…             677
#> # … with 49 more rows, and 5 more variables: delta.time.mins <dbl>,
#> #   delta.time.hours <dbl>, delta.time.days <dbl>, Directory <list>,
#> #   FileName <list>

# specify column to pass station names
get_record_table(camtrapdp,
    stationCol = "location_id",
    minDeltaTime = 20,
    deltaTimeComparedTo = "lastRecord")

#> Number of not independent observations to be removed: 82

#> # A tibble: 185 × 11
#>    Station        Species   DateTimeOriginal    Date       Time  delta.time.secs
#>    <chr>          <chr>     <dttm>              <date>     <chr>           <dbl>
#>  1 56128022-6061… Anas pla… 2020-06-03 20:04:33 2020-06-03 20:0…               0
#>  2 56128022-6061… Anas pla… 2020-06-04 02:10:17 2020-06-04 02:1…           21944
#>  3 56128022-6061… Anas pla… 2020-06-04 03:36:22 2020-06-04 03:3…            5165
#>  4 56128022-6061… Anas pla… 2020-06-06 01:51:53 2020-06-06 01:5…          166531
#>  5 7c989f1e-4fcb… Anas pla… 2020-06-05 20:24:17 2020-06-05 20:2…               0
#>  6 91d9abdd-da56… Anas pla… 2020-06-03 20:10:18 2020-06-03 20:1…               0
#>  7 91d9abdd-da56… Anas pla… 2020-06-03 20:57:50 2020-06-03 20:5…            2852
#>  8 91d9abdd-da56… Anas pla… 2020-06-03 23:19:10 2020-06-03 23:1…            8480
#>  9 91d9abdd-da56… Anas pla… 2020-06-04 21:05:24 2020-06-04 21:0…           78374
#> 10 91d9abdd-da56… Anas pla… 2020-06-05 02:01:54 2020-06-05 02:0…           17790
#> # … with 175 more rows, and 5 more variables: delta.time.mins <dbl>,
#> #   delta.time.hours <dbl>, delta.time.days <dbl>, Directory <list>,
#> #   FileName <list>