Get individuals

Gets the (unique) individuals from the observations of a Camera Trap Data Package object.

Usage

individuals(x)

Arguments

x

Camera Trap Data Package object, as returned by read_camtrapdp().

Value

A tibble::tibble() data frame with the individuals that have an individualID, containing the following columns:

individualID

scientificName

lifeStage

sex

See also

Examples

x <- example_dataset()
individuals(x)
#> # A tibble: 0 × 4
#> # ℹ 4 variables: individualID <chr>, scientificName <chr>, lifeStage <fct>,
#> #   sex <fct>