Visualize deployments features

This function visualizes deployments features such as number of species detected or number of observations on a dynamic map. The circle size and color are proportional to the mapped feature. Deployments without observations are shown as gray circles and a message is returned.

map_dep(
  datapkg,
  feature,
  cluster = TRUE,
  hover_columns = c("n", "deployment_id", "location_id", "location_name", "latitude",
    "longitude", "start", "end"),
  relative_color_scale = TRUE,
  max_color_scale = NULL,
  radius_range = c(10, 50)
)

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp(), i.e. a list containing three data.frames:

observations

deployments

multimedia

and a list with metadata: datapackage

feature
character, one of: - n_species - n_obs

cluster
a logical value indicating whether using the cluster option while visualizing maps. Default: TRUE

hover_columns
character with the name of the columns to use for showing location deployment information while hovering the mouse over. One or more from deployment columns. Use NULL to disable hovering. Default information:

n: number of species/observations (column created internally by a get_n_*() function)

start: start deployment

end: end deployment

deployment_id

location_id

location_name

latitude

longitude

See section Deployment of Camtrap DP standard for the full list of columns you can use

relative_color_scale
a logical indicating whether to use a relative color scale (TRUE) or an absolute scale (FALSE). If absolute scale is used, specify a valid max_color_scale

max_color_scale
a number indicating the max value used in the color scale

radius_range
a vector of length 2 containing the lower and upper limit of the circle radius. The lower value is used for deployments with zero observations/identified species. The upper value for the deployment(s) with the highest number of identified species/observations. Default: c(10, 50)

Value

a leaflet map

Details

Possible filtering about time period, ... (to be added)

Examples