Read camtrap-dp formatted data

This function reads camera trap data formatted following the Camera Trap Data Package (Camtrap DP) format.

read_camtrap_dp(path, media = TRUE)

Arguments

path
Path to the folder containing the camera trap data files.

media
If TRUE, read media records into memory. If FALSE, ignore media file to speed up reading larger Camtrap DP packages.

Value

A list containing three (tibble) data.frames:

observations

deployments

media

and a list with metadata: datapackage

Details

Vernacular names are typically used while working with camera trap observations, so they are added to the observations as defined in the metadata (slot taxonomic), if present.

Examples

if (FALSE) {
library(here)
# Read Camtrap DP package
camtrap_dp_dir <- here("inst", "extdata", "mica")
muskrat_coypu <- read_camtrap_dp(camtrap_dp_dir)

# Read Camtrap DP package and ignore media file
muskrat_coypu <- read_camtrap_dp(camtrap_dp_dir, media = FALSE)
}