Get taxa

Gets the (unique) scientific names and associated taxonomic information from the observations of a Camera Trap Data Package object.

Usage

taxa(x)

Arguments

x

Camera Trap Data Package object, as returned by read_camtrapdp().

Value

tibble() data frame with the taxonomic information, containing at least a scientificName column.

See also

Examples

x <- example_dataset()
taxa(x)
#> # A tibble: 10 × 5
#>    scientificName     taxonID  taxonRank vernacularNames.eng vernacularNames.nld
#>    <chr>              <chr>    <chr>     <chr>               <chr>              
#>  1 Anas platyrhynchos https:/… species   mallard             wilde eend         
#>  2 Rattus norvegicus  https:/… species   brown rat           bruine rat         
#>  3 Ardea cinerea      https:/… species   grey heron          blauwe reiger      
#>  4 Homo sapiens       https:/… species   human               mens               
#>  5 Anas strepera      https:/… species   gadwall             krakeend           
#>  6 Aves               https:/… class     bird sp.            vogel              
#>  7 Mustela putorius   https:/… species   European polecat    bunzing            
#>  8 Vulpes vulpes      https:/… species   red fox             vos                
#>  9 Martes foina       https:/… species   beech marten        steenmarter        
#> 10 Ardea              https:/… genus     great herons        reigers