Intermediate function to apply filter predicates on a data.frame

This function is used internally by all the get_*() functions to filter on deployments.

apply_filter_predicate(df, verbose, ...)

Arguments

df
data.frame we want to apply filter(s) expression(s)

verbose
Show (TRUE) or not (FALSE) the filter predicate expression

...
filter predicates to apply to df

Value

a data.frame

Examples

# and
apply_filter_predicate(camtrapdp$deployments,
                       verbose = TRUE,
                       pred_gte("latitude", 51.28),
                       pred_lt("longitude", 3.56))

#> df %>% filter((latitude >= 51.28) & (longitude < 3.56))

#> # A tibble: 0 x 16
#> # … with 16 variables: deployment_id <chr>, location_id <chr>,
#> #   location_name <chr>, longitude <dbl>, latitude <dbl>, start <dttm>,
#> #   end <dttm>, setup_by <chr>, camera_id <lgl>, camera_model <lgl>,
#> #   camera_interval <dbl>, camera_height <dbl>, bait_use <lgl>,
#> #   feature_type <lgl>, tags <chr>, comments <lgl>

# equivalent of
apply_filter_predicate(camtrapdp$deployments,
                       verbose = TRUE,
                       pred_and(pred_gte("latitude", 51.28),
                                pred_lt("longitude", 3.56)))

#> df %>% filter(((latitude >= 51.28) & (longitude < 3.56)))

#> # A tibble: 0 x 16
#> # … with 16 variables: deployment_id <chr>, location_id <chr>,
#> #   location_name <chr>, longitude <dbl>, latitude <dbl>, start <dttm>,
#> #   end <dttm>, setup_by <chr>, camera_id <lgl>, camera_model <lgl>,
#> #   camera_interval <dbl>, camera_height <dbl>, bait_use <lgl>,
#> #   feature_type <lgl>, tags <chr>, comments <lgl>

# or
apply_filter_predicate(camtrapdp$deployments,
                       verbose = TRUE,
                       pred_or(pred_gte("latitude", 51.28),
                                pred_lt("longitude", 3.56)))

#> df %>% filter(((latitude >= 51.28) | (longitude < 3.56)))

#> # A tibble: 2 x 16
#>   deployment_id location_id location_name longitude latitude start              
#>   <chr>         <chr>       <chr>             <dbl>    <dbl> <dttm>             
#> 1 3b94f935-020… 9a72bd3a-2… B_ML_val 03_…      3.55     51.3 2020-06-03 20:19:10
#> 2 6b1199ac-e6e… 91d9abdd-d… B_ML_val 05_…      3.57     51.3 2020-06-03 20:10:18
#> # … with 10 more variables: end <dttm>, setup_by <chr>, camera_id <lgl>,
#> #   camera_model <lgl>, camera_interval <dbl>, camera_height <dbl>,
#> #   bait_use <lgl>, feature_type <lgl>, tags <chr>, comments <lgl>