Get effort

Function to get the effort (deployment duration) per deployment.

get_effort(datapkg, ..., unit = "hour")

Arguments

datapkg

a camera trap data package object, as returned by read_camtrap_dp().

...

filter predicates

unit

time unit to use while returning deployment effort (duration). One of:

second

minute

hour

day

month

year

Value

a tibble (data.frame) with the following columns:

deploymentID: deployment unique identifier.

effort: effort expressed in the unit passed by argument unit.

unit: the unit used to express the effort. One of the values available for argument unit.

effort_duration: a duration object (duration is a class from lubridate package).

See also

Examples

# efforts expressed in hours
get_effort(mica)
#> # A tibble: 4 × 4
#>   deploymentID                         effort unit  effort_duration       
#>   <chr>                                 <dbl> <chr> <Duration>            
#> 1 29b7d356-4bb4-4ec4-b792-2af5cc32efa8   239. hour  859859s (~1.42 weeks) 
#> 2 577b543a-2cf1-4b23-b6d2-cda7e2eac372   219. hour  786802s (~1.3 weeks)  
#> 3 62c200a9-0e03-4495-bcd8-032944f6f5a1   529. hour  1903602s (~3.15 weeks)
#> 4 7ca633fa-64f8-4cfc-a628-6b0c419056d7   335. hour  1204929s (~1.99 weeks)

# effort expressed as days
get_effort(mica, unit = "day")
#> # A tibble: 4 × 4
#>   deploymentID                         effort unit  effort_duration       
#>   <chr>                                 <dbl> <chr> <Duration>            
#> 1 29b7d356-4bb4-4ec4-b792-2af5cc32efa8   9.95 day   859859s (~1.42 weeks) 
#> 2 577b543a-2cf1-4b23-b6d2-cda7e2eac372   9.11 day   786802s (~1.3 weeks)  
#> 3 62c200a9-0e03-4495-bcd8-032944f6f5a1  22.0  day   1903602s (~3.15 weeks)
#> 4 7ca633fa-64f8-4cfc-a628-6b0c419056d7  13.9  day   1204929s (~1.99 weeks)