Visualize deployments features

This function visualizes deployments features such as number of detected species, number of observations and RAI on a dynamic map. The circle size and color are proportional to the mapped feature. Deployments without observations are shown as gray circles and a message is returned.

map_dep(
  datapkg,
  feature,
  ...,
  species = NULL,
  sex = NULL,
  age = NULL,
  effort_unit = NULL,
  cluster = TRUE,
  hover_columns = c("n", "species", "deployment_id", "location_id", "location_name",
    "latitude", "longitude", "start", "end"),
  relative_scale = TRUE,
  max_scale = NULL,
  radius_range = c(10, 50)
)

Arguments

datapkg
a camera trap data package object, as returned by read_camtrap_dp(), i.e. a list containing three data.frames:

observations 2. deployments 3. multimedia

and a list with metadata: datapackage

feature
character, deployment feature to visualize. One of:

n_species: number of identified species

n_obs: number of observations

rai: Relative Abundance Index

effort: effort (duration) of the deployment

...
filter predicates for subsetting deployments

species
a character with a scientific name. Required for rai, optional for n_obs. Default: NULL

sex
a character defining the sex class to filter on, e.g. "female". If NULL, default, all observations of all sex classes are taken into account. Optional argument for n_obs

age
a character vector defining the age class to filter on, e.g. "adult" or c("subadult", "adult"). If NULL, default, all observations of all age classes are taken into account. Optional argument for n_obs

effort_unit
time unit to use while visualizing deployment effort (duration). One of:

second

minute

hour

day

month

year

NULL (default) duration objects (e.g. 2594308s (~4.29 weeks)) are shown while hovering and seconds shown in legend

cluster
a logical value indicating whether using the cluster option while visualizing maps. Default: TRUE

hover_columns
character with the name of the columns to use for showing location deployment information while hovering the mouse over. One or more from deployment columns. Use NULL to disable hovering. Default information:

n: number of species, number of observations, RAI or effort (column created internally by a get_*() function)

species: species name(s)

start: start deployment

end: end deployment -

deployment_id deployment unique identifier

location_id location unique identifier

location_name location name

latitude:

longitude:

See section Deployment of Camtrap DP standard for the full list of columns you can use

relative_scale
a logical indicating whether to use a relative color and radius scale (TRUE) or an absolute scale (FALSE). If absolute scale is used, specify a valid max_scale

max_scale
a number indicating the max value used to map color and radius

radius_range
a vector of length 2 containing the lower and upper limit of the circle radius. The lower value is used for deployments with zero feature value, i.e. no observations, no identified species, zero RAI or zero effort. The upper value is used for the deployment(s) with the highest feature value (relative_scale = TRUE) or max_scale (relative_scale = FALSE). Default: c(10, 50)

Value

a leaflet map

See also

Examples